Move get_args from helpers.py to run.py
